reference/requirements.txt
# Core dependencies
flask>=2.0.0
flask-cors>=3.0.0
psutil==5.9.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rich==13.7.0
SQLAlchemy==2.0.27

# Wayland-specific tools
wl-clipboard>=2.0.0; sys_platform == 'linux'
wl-paste>=2.0.0; sys_platform == 'linux'
swaymsg>=1.0.0; sys_platform == 'linux'
swayidle>=1.0.0; sys_platform == 'linux'